Fix issue with fs on cloudflare for the /icons endpoint.
