+++ src/pages/api/[version]/icons/[iconSet].ts
@@ -6,31 +6,31 @@ import { getIconSvgsForSet } from '../../../../utils/icons/reactIcons'
 /**
  * Prerender at build time so this doesn't run in the Cloudflare Worker.
  * getIconSvgsForSet() reads from @patternfly/react-icons/dist/static (Node fs).
- * Serves JSON of all icon SVGs for a set (e.g. /api/v5/icons/pf.json).
+ * Serves JSON of all icon SVGs for a set (e.g. /api/v6/icons/pf).
  */
 export const prerender = true
 
 export const getStaticPaths: GetStaticPaths = async () => {
   const versions = await getVersionsFromIndexFile()
   return versions.flatMap((version) => [
-    { params: { version, setId: 'pf' } },
+    { params: { version, iconSet: 'pf' } },
   ])
 }
 
 export const GET: APIRoute = async ({ params }) => {
-  const { version, setId } = params
+  const { version, iconSet } = params
   if (!version) {
     return createJsonResponse(
       { error: 'Version parameter is required' },
       400,
     )
   }
-  if (!setId) {
-    return createJsonResponse({ error: 'Set ID is required' }, 400)
+  if (!iconSet) {
+    return createJsonResponse({ error: 'Icon set is required' }, 400)
   }
 
   try {
-    const svgs = await getIconSvgsForSet(setId)
+    const svgs = await getIconSvgsForSet(iconSet)
     return createJsonResponse(svgs)
   } catch (error) {
     const details = error instanceof Error ? error.message : String(error)
