+++ src/utils/icons/fetch.ts
@@ -1,4 +1,4 @@
-import type { IconMetadata } from './reactIcons'
+import type { IconMetadata } from './icons'
 
 export interface IconsIndex {
   icons: IconMetadata[]
@@ -41,7 +41,7 @@ export async function fetchIconSvgs(
   version: string,
   setId: string,
 ): Promise<Record<string, string> | null> {
-  const iconsSvgsUrl = new URL(`/api/${version}/icons/${setId}.json`, url.origin)
+  const iconsSvgsUrl = new URL(`/api/${version}/icons/${setId}`, url.origin)
   const response = assetsFetch
     ? await assetsFetch(new Request(iconsSvgsUrl))
     : await fetch(iconsSvgsUrl)
