+++ src/utils/icons/icons.ts
@@ -1,0 +1,7 @@
+export interface IconMetadata {
+    name: string
+    reactName: string
+    usage: string
+  }
+  
+  export const PF_ICONS_SET_ID = 'pf'
